+++ point2d.h
@@ -11,10 +11,10 @@ Atomic class for 2D points, mostly used to build trajectory so we don't have to 
 struct Point2d {
     Point2d();
     Point2d(float _x, float _y);
-    Point2d(Eigen::Vector2f _v);
+    Point2d(Eigen::Vector2d _v);
 
     float x; float y;
-    Eigen::Vector2f position;
+    Eigen::Vector2d position;
 
     void randomize(const float &lower_bound, const float &upper_bound);
     void print() const;
